+++ packages/thumbnails/test/thumbnailDecorator.spec.ts
@@ -56,8 +56,7 @@ describe("thumbnailDecorator", () => {
     const decorator = thumbnailDecorator({ size: "7" })
 
     const shopifyUrl = "https://cdn.shopify.com/s/files/1/0097/5821/2174/files/clothing-red.jpg?v=12345"
-    const modifiedUrl =
-      "https://cdn.shopify.com/s/files/1/0097/5821/2174/files/clothing-red_200x200_crop_center.jpg?v=12345"
+    const modifiedUrl = "https://cdn.shopify.com/s/files/1/0097/5821/2174/files/clothing-red_200x200.jpg?v=12345"
 
     const result = decorator({
       imageUrl: shopifyUrl